+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/TaskOutputOriginFixture.groovy
@@ -20,7 +20,7 @@ import groovy.json.JsonSlurper
 import org.gradle.api.internal.tasks.OriginTaskExecutionMetadata
 import org.gradle.integtests.fixtures.executer.GradleExecuter
 import org.gradle.integtests.fixtures.executer.UserInitScriptExecuterFixture
-import org.gradle.internal.execution.ExecuteTaskBuildOperationType
+import org.gradle.api.internal.tasks.execution.ExecuteTaskBuildOperationType
 import org.gradle.internal.id.UniqueId
 import org.gradle.internal.operations.BuildOperationDescriptor
 import org.gradle.internal.operations.BuildOperationListener
